+++ src/accelerator/ogl/image/image_kernel.h
@@ -42,7 +42,6 @@ enum class keyer
     additive,
 };
 
-// TODO QB: add boolean flag straight_alpha_output
 struct draw_params final
 {
     core::pixel_format_desc                     pix_desc = core::pixel_format_desc(core::pixel_format::invalid);
@@ -55,6 +54,7 @@ struct draw_params final
     std::shared_ptr<class texture>              local_key;
     std::shared_ptr<class texture>              layer_key;
     double                                      aspect_ratio = 1.0;
+    bool                                        straight_alpha_output = false;
     int                                         target_width;
     int                                         target_height;
 };